refactor(lora): share full dtype-aware CPU delta; hybrid CPU LoRA now supports F16/BF16/Q8_0 (#82 follow-up)
Extract TransformerModel.ApplyLoraDelta's body verbatim into a new shared
static LoraProjection.Apply (DotLLM.Models) callable by both TransformerModel
and HybridTransformerModel. TransformerModel.ApplyLoraDelta now delegates
(behavior-preserving, early null-guard retained). HybridTransformerModel's
CPU-phase ApplyLoraDeltaCpu re-points to the shared helper, dropping its
F32-only path and the NotSupportedException for non-F32 adapter weights.
Hybrid CPU-resident layers now get full F32/F16/BF16/Q8_0 dtype support plus
the Phase-4d.6 outer-product stage-2 fast path.
